Refactor address and wallet view into common component
